♻️  refactor(#85): 대시보드 헤더 컴포넌트 파일 분리
